Adds ExtraFieldPropagation.Customizer to implement secondary sampling
Currently, we have a test `SecondarySampling` which is complex to
implement as it assumes conditional decisions will result in use of
different reporting libraries. Ex normal zipkin for sampled data, and
a different format for secondary sinks.

After a chat with Nara, we realized that data routing logic is cheapest
to address outside the process. In other words, if a secondary sampling
decision says "record" but only for a part of the network, this data
for that subtree will still end up posting to the same place.

With this in mind, secondary sampling can be far simpler, driven mainly
by decisions made on "extra fields". This change adds two things to
allow this to happen.

* `ExtraFieldPropagation.Customizer` which allows you to set a secondary decision (ex via sampledLocal)
* `Tracing.Builder.alwaysReportSpans()` which allows the normal span reporter to get everything recorded.

This PR neetds a little more work as the `SecondarySampling` test needs
to be redone to prove it makes things simpler.

Fixes #955